pyproject.toml
[tool.poetry]
name = "l5-bot"
version = "0.1.0"
description = ""
authors = ["Denis King <newmancu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^21.6"
loguru = "^0.7.2"
pydantic = "^2.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
